Move FlowSafeSubscriber and flowSafeSubscribe under their own package
